Minor code style changes. Unnecessary comments removed
