+++ packages/itwin/ec3-widget/src/components/Templates.scss
@@ -4,7 +4,7 @@
 *--------------------------------------------------------------------------------------------*/
 @import "~@itwin/itwinui-css/scss/variables";
 
-.e_c_3 {
+.ec3 {
   &-scrollable-table {
     overflow-y: scroll;
     max-height: 90%;
