+++ packages/itwin/ec3-widget/src/index.d.ts
@@ -2,4 +2,4 @@
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
-/// <reference types="@bentley/react-scripts" />
+declare module "*.png";
