+++ rush.json
@@ -10,7 +10,9 @@
   "projectFolderMaxDepth": 3,
   "ensureConsistentVersions": false,
   "gitPolicy": {
-    "allowedEmailRegExps": ["[^@]+@users\\.noreply\\.github\\.com"],
+    "allowedEmailRegExps": [
+      "[^@]+@users\\.noreply\\.github\\.com"
+    ],
     "sampleEmail": "mrexample@users.noreply.github.com"
   },
   "repository": {
@@ -68,7 +70,7 @@
       "versionPolicyName": "infrastructure-iot"
     },
     {
-      "packageName": "@itwin/ec3-react",
+      "packageName": "@itwin/ec3-widget-react",
       "projectFolder": "packages/itwin/ec3-widget",
       "versionPolicyName": "ec3-widget"
     }
